do_function/monitor_snapshots/packages/monitor-snapshot/main/requirements.txt
slackclient==2.9.4
requests==2.26.0